requirements.txt
requests
asgiref==3.5.2
Django==3.2
sqlparse==0.4.2

djangorestframework==3.12.4
pytz==2021.1

drf-yasg==1.21.3

djoser==2.1.0
djangorestframework_simplejwt
PyJWT==1.7.1
django-debug-toolbar
django-cors-headers==2.4.0
django-cors-middleware==1.3.1


